requirements.txt
requests==2.31.0
pandas==2.2.1
python-dotenv==1.0.1
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
loguru==0.7.2 